Unify e2e tests into one file
